+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/mapper/FollowMapper.java
@@ -9,17 +9,17 @@ import java.util.List;
 public interface FollowMapper {
 
     @Select("SELECT * FROM `follow` WHERE follower_id = #{followerId} and followee_id = #{followeeId}")
-    FollowDO get(int followerId, int followeeId);
+    FollowDO get(long followerId, long followeeId);
 
     @Select("SELECT * FROM `follow` " +
             "WHERE follower_id = #{followerId} and created_at < #{createdAt} " +
             "order by created_at DESC " +
             "LIMIT #{limit} ")
-    List<FollowDO> getList(int followerId, LocalDateTime createdAt, int limit);
+    List<FollowDO> getList(long followerId, LocalDateTime createdAt, int limit);
 
     @Insert("INSERT INTO follow(follower_id, followee_id) VALUES (#{followerId}, #{followeeId})")
-    int insert(int followerId, int followeeId);
+    int insert(long followerId, long followeeId);
 
     @Update("DELETE FROM `follow` WHERE follower_id = #{followerId} and followee_id = #{followeeId}")
-    void delete(int followerId, int followeeId);
+    void delete(long followerId, long followeeId);
 }
