+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/mapper/ProfessionMapper.java
@@ -10,22 +10,22 @@ import java.util.Map;
 public interface ProfessionMapper {
 
     @Select("SELECT * FROM profession WHERE id = #{id}")
-    ProfessionDO get(int id);
+    ProfessionDO get(long id);
 
     @Select("SELECT * FROM profession ORDER BY id DESC LIMIT #{size} OFFSET #{offset}")
     List<ProfessionDO> listByPage(int offset, int size);
 
     @Select("SELECT * FROM profession WHERE state = #{state} AND id > #{lastId} ORDER BY id ASC LIMIT 20")
-    List<ProfessionDO> listByStateAndLastId(String state, int lastId);
+    List<ProfessionDO> listByStateAndLastId(String state, long lastId);
 
     @Select("SELECT * FROM profession WHERE main_category = #{mainCategory} AND state = 'APPROVED' AND id > #{lastId} ORDER BY id ASC LIMIT 20")
-    List<ProfessionDO> listByMainCategoryAndLastId(int mainCategory, int lastId);
+    List<ProfessionDO> listByMainCategoryAndLastId(int mainCategory, long lastId);
 
     @Select("SELECT * FROM profession WHERE sub_category = #{subCategory} AND state = 'APPROVED' AND id > #{lastId} ORDER BY id ASC LIMIT 20")
-    List<ProfessionDO> listBySubCategoryAndLastId(int subCategory, int lastId);
+    List<ProfessionDO> listBySubCategoryAndLastId(int subCategory, long lastId);
 
     @Select("SELECT * FROM profession WHERE main_category = #{mainCategory} AND sub_category = #{subCategory} AND state = 'APPROVED' AND id > #{lastId} ORDER BY id ASC LIMIT 20")
-    List<ProfessionDO> listByMainCategoryAndSubCategoryAndLastId(int mainCategory, int subCategory, int lastId);
+    List<ProfessionDO> listByMainCategoryAndSubCategoryAndLastId(int mainCategory, int subCategory, long lastId);
 
     @Select("SELECT * FROM profession WHERE INSTR(name, #{name}) > 0 limit 20")
     List<ProfessionDO> searchByName(String name);
@@ -55,17 +55,17 @@ public interface ProfessionMapper {
     void update(ProfessionDO professionDO);
 
     @Update("UPDATE profession SET state = 'APPROVED', updated_at = CURRENT_TIMESTAMP WHERE id = #{id}")
-    int approve(int id);
+    int approve(long id);
 
     @Update("UPDATE profession SET state = 'REJECTED', rejected_reason = #{rejectedReason}, updated_at = CURRENT_TIMESTAMP " +
             "WHERE id = #{id}")
-    int reject(int id, String rejectedReason);
+    int reject(long id, String rejectedReason);
 
     @Update("UPDATE profession SET state = 'BLOCKED', updated_at = CURRENT_TIMESTAMP WHERE id = #{id}")
-    int block(int id);
+    int block(long id);
 
     @Delete("DELETE FROM profession WHERE id = #{id}")
-    void delete(int id);
+    void delete(long id);
     
     // 平台统计相关方法
     @Select("SELECT COUNT(*) FROM profession WHERE state = 'APPROVED'")
