+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/mapper/RoadmapMapper.java
@@ -8,20 +8,20 @@ import java.util.List;
 public interface RoadmapMapper {
 
     @Select("SELECT * FROM roadmap WHERE id = #{id}")
-    RoadmapDO get(int id);
+    RoadmapDO get(long id);
 
     @Select("SELECT * FROM roadmap ORDER BY created_at DESC LIMIT #{offset}, #{limit}")
     List<RoadmapDO> getList(int offset, int limit);
 
     @Select("SELECT * FROM roadmap WHERE profession_id = #{professionId} ORDER BY created_at DESC LIMIT #{offset}, #{limit}")
-    List<RoadmapDO> getListByProfession(int professionId, int offset, int limit);
+    List<RoadmapDO> getListByProfession(long professionId, int offset, int limit);
 
     @Select({"<script>",
              "SELECT * FROM roadmap WHERE id IN ",
              "<foreach item='id' collection='ids' open='(' separator=',' close=')'>#{id}</foreach>",
              " ORDER BY created_at DESC",
              "</script>"})
-    List<RoadmapDO> getByIds(List<Integer> ids);
+    List<RoadmapDO> getByIds(List<Long> ids);
 
     @Select({"<script>",
              "SELECT * FROM roadmap WHERE profession_id = #{professionId}",
@@ -31,7 +31,7 @@ public interface RoadmapMapper {
              "</if>",
              " ORDER BY created_at DESC LIMIT #{offset}, #{limit}",
              "</script>"})
-    List<RoadmapDO> getListByProfessionExcluding(int professionId, int offset, int limit, List<Integer> excludeIds);
+    List<RoadmapDO> getListByProfessionExcluding(long professionId, int offset, int limit, List<Long> excludeIds);
 
     @Select({"<script>",
              "SELECT * FROM roadmap WHERE profession_id = #{professionId} AND id &lt; #{lastId}",
@@ -42,10 +42,10 @@ public interface RoadmapMapper {
              " ORDER BY id DESC LIMIT #{limit}",
              "</script>"})
     List<RoadmapDO> getListByProfessionAfterIdExcluding(
-            int professionId, int lastId, int limit, List<Integer> excludeIds);
+            long professionId, long lastId, int limit, List<Long> excludeIds);
 
     @Select("SELECT * FROM roadmap WHERE creator_id = #{creatorId} ORDER BY created_at DESC LIMIT #{offset}, #{limit}")
-    List<RoadmapDO> getListByCreator(int creatorId, int offset, int limit);
+    List<RoadmapDO> getListByCreator(long creatorId, int offset, int limit);
 
     @Select("SELECT * FROM roadmap WHERE content_hash = #{contentHash}")
     List<RoadmapDO> getByContentHash(String contentHash);
@@ -59,10 +59,10 @@ public interface RoadmapMapper {
     void update(RoadmapDO roadmapDO);
 
     @Update("UPDATE roadmap SET vote = vote + #{delta} WHERE id = #{id} AND (vote + #{delta}) >= 0")
-    int updateVoteCount(int id, int delta);
+    int updateVoteCount(long id, int delta);
 
     @Update("UPDATE roadmap SET score = #{score}, score_calculated_at = NOW() WHERE id = #{id}")
-    int updateScore(int id, Double score);
+    int updateScore(long id, Double score);
 
     @Select("SELECT * FROM roadmap WHERE id IN " +
             "(SELECT DISTINCT post_id FROM upvote WHERE post_type = 'roadmap') " +
@@ -78,7 +78,7 @@ public interface RoadmapMapper {
              " ORDER BY score DESC, id DESC LIMIT #{offset}, #{limit}",
              "</script>"})
     List<RoadmapDO> getListByProfessionExcludingOrderByScore(
-            int professionId, int offset, int limit, List<Integer> excludeIds);
+            long professionId, int offset, int limit, List<Long> excludeIds);
 
     @Select({"<script>",
              "SELECT * FROM roadmap WHERE profession_id = #{professionId} AND ",
@@ -90,7 +90,7 @@ public interface RoadmapMapper {
              " ORDER BY score DESC, id DESC LIMIT #{limit}",
              "</script>"})
     List<RoadmapDO> getListByProfessionAfterScoreExcluding(
-            int professionId, Double lastScore, int lastId, int limit, List<Integer> excludeIds);
+            long professionId, Double lastScore, long lastId, int limit, List<Long> excludeIds);
     
     // 平台统计相关方法
     @Select("SELECT COUNT(*) FROM roadmap WHERE vote >= 0")
