+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/mapper/UpvoteMapper.java
@@ -8,12 +8,12 @@ import java.util.List;
 public interface UpvoteMapper {
 
     @Select("SELECT * FROM upvote where user_id = #{userId} and object_id = #{objectId} and object_type = #{objectType}")
-    UpvoteDO get(int userId, int objectId, int objectType);
+    UpvoteDO get(long userId, long objectId, int objectType);
 
     @Select({"<script>SELECT object_id, type FROM upvote where user_id = #{userId} and object_type = #{objectType} and object_id in " +
             "<foreach item='id' collection='ids' open='(' separator=', ' close=')'>#{id}</foreach>" +
             "</script>"})
-    List<UpvoteDO> getList(int userId, List<Integer> ids, int objectType);
+    List<UpvoteDO> getList(long userId, List<Long> ids, int objectType);
 
     @Insert("INSERT INTO upvote(user_id, object_id, object_type, type) VALUES (#{userId}, #{objectId}, #{objectType}, #{type})")
     @Options(useGeneratedKeys = true, keyProperty = "id", keyColumn = "id")
@@ -23,5 +23,5 @@ public interface UpvoteMapper {
     void update(UpvoteDO upvoteDO);
 
     @Delete("DELETE FROM upvote where id = #{id}")
-    void delete(int id);
+    void delete(long id);
 }
